refactor(libs/common): formatFileSize と formatPrice を集約 (#2994)
- libs/common に format モジュールを新設し、formatFileSize と formatPrice を提供
- codec-converter/core の formatFileSize を削除し、呼出側 (web/src/app/page.tsx,
  web/src/app/jobs/[jobId]/page.tsx) を @nagiyu/common 経由に切替
- stock-tracker/web の formatPrice を削除し、AlertSettingsModal を @nagiyu/common
  経由に切替
- codec-converter/web/src/app/page.tsx のローカル formatFileSize（toFixed(2)）を
  libs/common 版（toFixed(1)）に統一。アップロードファイルサイズ表示で小数桁が
  2 桁から 1 桁に変わるが UX 影響は軽微
- candidates.md を 4 件構成に更新。G は実装前確認で「既に共通化済み」と判明した
  ためスキップ、E は縮小スコープ実施として記録
