+++ libs/common/src/format/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Format Module
+ *
+ * 数値・ファイルサイズなど、サービス横断で再利用される表示用フォーマッタを提供する。
+ */
+
+export { formatFileSize } from './file-size.js';
+export { formatPrice } from './price.js';
