+++ services/codec-converter/core/src/format.ts
@@ -1,16 +1,4 @@
 /**
- * ファイルサイズを人間が読める形式にフォーマット
- * @param bytes ファイルサイズ（バイト）
- * @returns フォーマットされた文字列（KB または MB）
- */
-export function formatFileSize(bytes: number): string {
-  if (bytes < 1024 * 1024) {
-    return `${(bytes / 1024).toFixed(1)} KB`;
-  }
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-}
-
-/**
  * Unix timestampを日時文字列にフォーマット
  * @param timestamp Unix timestamp（秒）
  * @returns 日本語ロケールでフォーマットされた日時文字列
